requirements.txt
streamlit
sentence-transformers==2.6.1
faiss-cpu
numpy
requests
textblob
PyPDF2
